Major update

+++ backend/controllers/postController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler');
 const Post = require('../models/postModel');
 const Comment = require('../models/commentModel');
+const User = require('../models/userModel');
 const { v4: uuidv4 } = require('uuid');
 const sharp = require('sharp');
 const fs = require('fs');
@@ -9,9 +10,31 @@ const fs = require('fs');
 // @route   GET /api/posts
 // @access  Private
 const getPosts = asyncHandler(async (req, res) => {
-	const posts = await Post.find({ user: req.user.id });
-
-	res.status(200).json(posts);
+	await Post.find({ user: req.user.id });
+	Post.aggregate([
+		{
+			$lookup: {
+				from: 'users', // collection name in db
+				localField: 'user',
+				foreignField: '_id',
+				as: 'userData',
+			},
+		},
+	]).exec(function (err, postData) {
+		var userPostData = postData.map((data) => {
+			return data.userData.map((obj) => ({
+				...data,
+				userData: {
+					_id: obj._id,
+					email: obj.email,
+					firstName: obj.firstName,
+					lastName: obj.lastName,
+					photo: obj.photo,
+				},
+			}));
+		});
+		res.status(200).json(userPostData.map((data) => data[0]));
+	});
 });
 
 // @desc    Set post
@@ -32,15 +55,24 @@ const setPost = asyncHandler(async (req, res) => {
 			.toFile(`frontend/public/uploads/${fileName}`);
 	}
 
-	const post = await Post.create({
+	var post = await Post.create({
 		user: req.user.id,
 		text: req.body.text,
 		photo: fileName ? fileName : null,
 		audience: req.body.audience,
-		likes: {
-			userId: req.body.userId,
-		},
 	});
+
+	const user = await User.findById(post.user);
+
+	post = {
+		...post._doc,
+		userData: {
+			email: user.email,
+			firstName: user.firstName,
+			lastName: user.lastName,
+			photo: user.photo,
+		},
+	};
 
 	res.status(200).json(post);
 });
@@ -49,17 +81,17 @@ const setPost = asyncHandler(async (req, res) => {
 // @route   PUT /api/posts/:id
 // @access  Private
 const updatePost = asyncHandler(async (req, res) => {
+	// Check for user
+	if (!req.user) {
+		res.status(401);
+		throw new Error('User not found');
+	}
+
 	const post = await Post.findById(req.params.id);
 
 	if (!post) {
 		res.status(400);
 		throw new Error('Post not found');
-	}
-
-	// Check for user
-	if (!req.user) {
-		res.status(401);
-		throw new Error('User not found');
 	}
 
 	// Make sure the logged in user matches the post user
@@ -109,6 +141,12 @@ const updatePost = asyncHandler(async (req, res) => {
 // @route   DELETE /api/posts/:id
 // @access  Private
 const deletePost = asyncHandler(async (req, res) => {
+	// Check for user
+	if (!req.user) {
+		res.status(401);
+		throw new Error('User not found');
+	}
+
 	const post = await Post.findById(req.params.id);
 
 	await Comment.deleteMany({ postId: req.params.id });
@@ -116,12 +154,6 @@ const deletePost = asyncHandler(async (req, res) => {
 	if (!post) {
 		res.status(400);
 		throw new Error('Post not found');
-	}
-
-	// Check for user
-	if (!req.user) {
-		res.status(401);
-		throw new Error('User not found');
 	}
 
 	// Make sure the logged in user matches the post user
@@ -146,9 +178,15 @@ const deletePost = asyncHandler(async (req, res) => {
 });
 
 // @desc    Like post
-// @route   PUT /api/posts/:id
+// @route   PUT /api/posts/:id/likes
 // @access  Private
 const likePost = asyncHandler(async (req, res) => {
+	// Check for user
+	if (!req.user) {
+		res.status(401);
+		throw new Error('User not found');
+	}
+
 	const post = await Post.findById(req.params.id);
 
 	if (!post) {
@@ -156,13 +194,10 @@ const likePost = asyncHandler(async (req, res) => {
 		throw new Error('Post not found');
 	}
 
-	// Check for user
-	if (!req.user) {
-		res.status(401);
-		throw new Error('User not found');
-	}
-
-	const postLikes = await Post.find({ 'likes.userId': req.body.userId, '_id': req.body.postId });
+	const postLikes = await Post.find({
+		'likes.userId': req.body.userId,
+		_id: req.body.postId,
+	});
 
 	var likeResponse;
 
@@ -195,10 +230,64 @@ const likePost = asyncHandler(async (req, res) => {
 	res.status(200).json(likeResponse);
 });
 
+// @desc    Save post
+// @route   PUT /api/posts/:id/saves
+// @access  Private
+const savePost = asyncHandler(async (req, res) => {
+	// Check for user
+	if (!req.user) {
+		res.status(401);
+		throw new Error('User not found');
+	}
+
+	const post = await Post.findById(req.params.id);
+
+	if (!post) {
+		res.status(400);
+		throw new Error('Post not found');
+	}
+
+	const postSaves = await Post.find({
+		'saves.userId': req.body.userId,
+		_id: req.body.postId,
+	});
+
+	var saveResponse;
+
+	if (postSaves.length > 0) {
+		saveResponse = await Post.findByIdAndUpdate(
+			req.params.id,
+			{
+				$pull: {
+					'saves.userId': req.body.userId,
+				},
+			},
+			{
+				new: true,
+			}
+		);
+	} else {
+		saveResponse = await Post.findByIdAndUpdate(
+			req.params.id,
+			{
+				$push: {
+					'saves.userId': req.body.userId,
+				},
+			},
+			{
+				new: true,
+			}
+		);
+	}
+
+	res.status(200).json(saveResponse);
+});
+
 module.exports = {
 	getPosts,
 	setPost,
 	updatePost,
 	deletePost,
 	likePost,
+	savePost,
 };
